+++ src/index.js
@@ -52,6 +52,10 @@ export {
 
 export { ouIdHelper } from './modules/ouIdHelper'
 
+// Modules: adapted ui layout
+
+export { getAdaptedUiLayoutByType } from './modules/getAdaptedUiLayoutByType'
+
 // Modules: relativeItems
 
 export { hasRelativeItems } from './modules/relativeItems'
@@ -73,7 +77,6 @@ export { layoutHasDataDimension } from './modules/layout/layoutHasDataDimension'
 export { layoutHasDimension } from './modules/layout/layoutHasDimension'
 export { layoutHasDynamicDimension } from './modules/layout/layoutHasDynamicDimension'
 export { layoutHasPeriodDimension } from './modules/layout/layoutHasPeriodDimension'
-export { layoutGetAdaptedLayoutByType } from './modules/layout/layoutGetAdaptedLayoutByType'
 
 export {
     AXIS,
